# api/openapi.json
{
  "components": {
    "schemas": {
...
      "CompareRequest": {
        "additionalProperties": false,
        "description": "The same product under every registered method.",
        "properties": {
          "convergence": {
            "default": true,
            "title": "Convergence",
            "type": "boolean"
          },
          "pricing": {
            "$ref": "#/components/schemas/PriceRequest"
          }
        },
        "required": [
          "pricing"
        ],
        "title": "CompareRequest",
        "type": "object"
      },
      "CompareResponse": {
        "additionalProperties": false,
        "description": "Methods side by side.",
        "properties": {
          "boundary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExerciseBoundary"
              },
              {
                "type": "null"
              }
            ]
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "european_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Same contract with European exercise (unit, closed form)",
            "title": "European Price"
          },
          "methods": {
            "items": {
              "$ref": "#/components/schemas/MethodComparison"
            },
            "title": "Methods",
            "type": "array"
          }
        },
        "required": [
          "currency",
          "methods"
        ],
        "title": "CompareResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Engine error surfaced to the client.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ExerciseBoundary": {
        "additionalProperties": false,
        "description": "PDE early-exercise boundary: exercise when spot is beyond ``spot`` at ``days``.",
        "properties": {
          "days": {
            "description": "Days from the valuation date",
            "items": {
              "type": "number"
            },
            "title": "Days",
            "type": "array"
          },
          "spot": {
            "items": {
              "type": "number"
            },
            "title": "Spot",
            "type": "array"
          }
        },
        "required": [
          "days",
          "spot"
        ],
        "title": "ExerciseBoundary",
        "type": "object"
      },
...
      "HeatmapRequest": {
        "additionalProperties": false,
        "description": "Full-revaluation value grid over spot and vol shocks.",
        "properties": {
          "horizon_days": {
            "default": 0,
            "minimum": 0.0,
            "title": "Horizon Days",
            "type": "integer"
          },
          "pricing": {
            "$ref": "#/components/schemas/PriceRequest"
          },
          "spot_range_pct": {
            "default": 20.0,
            "exclusiveMaximum": 100.0,
            "exclusiveMinimum": 0.0,
            "title": "Spot Range Pct",
            "type": "number"
          },
          "spot_steps": {
            "default": 21,
            "maximum": 101.0,
            "minimum": 3.0,
            "title": "Spot Steps",
            "type": "integer"
          },
          "vol_range_pts": {
            "default": 10.0,
            "exclusiveMinimum": 0.0,
            "maximum": 100.0,
            "title": "Vol Range Pts",
            "type": "number"
          },
          "vol_steps": {
            "default": 21,
            "maximum": 101.0,
            "minimum": 3.0,
            "title": "Vol Steps",
            "type": "integer"
          }
        },
        "required": [
          "pricing"
        ],
        "title": "HeatmapRequest",
        "type": "object"
      },
      "HeatmapResponse": {
        "additionalProperties": false,
        "description": "Position values on a vol \u00d7 spot grid (rows: vol shifts). ``null`` = shock out of domain.",
        "properties": {
          "base_value": {
            "description": "Current position value, currency",
            "title": "Base Value",
            "type": "number"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "horizon_days": {
            "title": "Horizon Days",
            "type": "integer"
          },
          "notional": {
            "description": "quantity \u00d7 current spot, currency",
            "title": "Notional",
            "type": "number"
          },
          "spot_shifts_pct": {
            "items": {
              "type": "number"
            },
            "title": "Spot Shifts Pct",
            "type": "array"
          },
          "values": {
            "items": {
              "items": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "title": "Values",
            "type": "array"
          },
          "vol_shifts_pts": {
            "items": {
              "type": "number"
            },
            "title": "Vol Shifts Pts",
            "type": "array"
          }
        },
        "required": [
          "currency",
          "spot_shifts_pct",
          "vol_shifts_pts",
          "horizon_days",
          "base_value",
          "notional",
          "values"
        ],
        "title": "HeatmapResponse",
        "type": "object"
      },
      "ImpliedVolRequest": {
        "additionalProperties": false,
        "description": "Implied vol from a unit price. ``market.vol`` is ignored.",
        "properties": {
          "instrument": {
            "discriminator": {
              "mapping": {
                "american": "#/components/schemas/AmericanOptionIn",
                "digital": "#/components/schemas/DigitalOptionIn",
                "european": "#/components/schemas/EuropeanOptionIn"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/EuropeanOptionIn"
              },
              {
                "$ref": "#/components/schemas/AmericanOptionIn"
              },
              {
                "$ref": "#/components/schemas/DigitalOptionIn"
              }
            ],
            "title": "Instrument"
          },
          "market": {
            "$ref": "#/components/schemas/MarketIn"
          },
          "model": {
            "$ref": "#/components/schemas/ModelIn",
            "default": {
              "dividend_treatment": "escrowed",
              "type": "bsm"
            }
          },
          "settings": {
            "$ref": "#/components/schemas/SettingsIn",
            "default": {
              "bumps": {
                "rate_abs": 0.0001,
                "spot_rel": 0.001,
                "time_days": 1,
                "vol_abs": 0.001
              },
              "digital": {
                "spread_width_rel": 0.01
              },
              "force_bump_greeks": false,
              "implied_vol": {
                "max_iterations": 2
              },
              "pde": {
                "n_std": 5.0,
                "penalty": 100000000.0,
                "penalty_max_iter": 25,
                "penalty_tol": 1e-12,
                "rannacher_steps": 2,
                "space_nodes": 800,
                "time_steps": 200
              },
              "scenario": {
                "pde_space_nodes": 200,
                "pde_time_steps": 100,
                "tree_steps": 101
              },
              "tree": {
                "steps": 401
              }
            }
          },
          "target_price": {
            "description": "Unit price, currency",
            "exclusiveMinimum": 0.0,
            "title": "Target Price",
            "type": "number"
          }
        },
        "required": [
          "instrument",
          "market",
          "target_price"
        ],
        "title": "ImpliedVolRequest",
        "type": "object"
      },
      "ImpliedVolResponse": {
        "additionalProperties": false,
        "description": "Implied vol and the quantities it was solved from.",
        "properties": {
          "discount_factor": {
            "title": "Discount Factor",
            "type": "number"
          },
          "forward": {
            "title": "Forward",
            "type": "number"
          },
          "implied_vol": {
            "description": "Decimal",
            "title": "Implied Vol",
            "type": "number"
          },
          "time_to_expiry": {
            "title": "Time To Expiry",
            "type": "number"
          }
        },
        "required": [
          "implied_vol",
          "forward",
          "discount_factor",
          "time_to_expiry"
        ],
        "title": "ImpliedVolResponse",
        "type": "object"
      },
...
      "PriceRequest": {
        "additionalProperties": false,
        "description": "Price one instrument.",
        "properties": {
          "instrument": {
            "discriminator": {
              "mapping": {
                "american": "#/components/schemas/AmericanOptionIn",
                "digital": "#/components/schemas/DigitalOptionIn",
                "european": "#/components/schemas/EuropeanOptionIn"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/EuropeanOptionIn"
              },
              {
                "$ref": "#/components/schemas/AmericanOptionIn"
              },
              {
                "$ref": "#/components/schemas/DigitalOptionIn"
              }
            ],
            "title": "Instrument"
          },
          "market": {
            "$ref": "#/components/schemas/MarketIn"
          },
          "method": {
            "default": "analytic",
            "title": "Method",
            "type": "string"
          },
          "model": {
            "$ref": "#/components/schemas/ModelIn",
            "default": {
              "dividend_treatment": "escrowed",
              "type": "bsm"
            }
          },
          "settings": {
            "$ref": "#/components/schemas/SettingsIn",
            "default": {
              "bumps": {
                "rate_abs": 0.0001,
                "spot_rel": 0.001,
                "time_days": 1,
                "vol_abs": 0.001
              },
              "digital": {
                "spread_width_rel": 0.01
              },
              "force_bump_greeks": false,
              "implied_vol": {
                "max_iterations": 2
              },
              "pde": {
                "n_std": 5.0,
                "penalty": 100000000.0,
                "penalty_max_iter": 25,
                "penalty_tol": 1e-12,
                "rannacher_steps": 2,
                "space_nodes": 800,
                "time_steps": 200
              },
              "scenario": {
                "pde_space_nodes": 200,
                "pde_time_steps": 100,
                "tree_steps": 101
              },
              "tree": {
                "steps": 401
              }
            }
          }
        },
        "required": [
          "instrument",
          "market"
        ],
        "title": "PriceRequest",
        "type": "object"
      },
      "PriceResponse": {
        "additionalProperties": false,
        "description": "Price, market-derived quantities, greeks and diagnostics.",
        "properties": {
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "diagnostics": {
            "$ref": "#/components/schemas/DiagnosticsOut"
          },
          "discount_factor": {
            "title": "Discount Factor",
            "type": "number"
          },
          "forward": {
            "description": "Model forward to expiry, including discrete dividends",
            "title": "Forward",
            "type": "number"
          },
          "greeks": {
            "$ref": "#/components/schemas/GreeksOut"
          },
          "notional": {
            "description": "quantity \u00d7 spot, currency",
            "title": "Notional",
            "type": "number"
          },
          "pct_notional": {
            "description": "100 \u00d7 unit price / spot, %",
            "title": "Pct Notional",
            "type": "number"
          },
          "position_value": {
            "description": "quantity \u00d7 unit price, currency",
            "title": "Position Value",
            "type": "number"
          },
          "price": {
            "description": "Unit price, currency",
            "title": "Price",
            "type": "number"
          },
          "time_to_expiry": {
            "description": "ACT/365F year fraction",
            "title": "Time To Expiry",
            "type": "number"
          }
        },
        "required": [
          "currency",
          "price",
          "position_value",
          "pct_notional",
          "notional",
          "forward",
          "discount_factor",
          "time_to_expiry",
          "greeks",
          "diagnostics"
        ],
        "title": "PriceResponse",
        "type": "object"
      },
      "ProfileRequest": {
        "additionalProperties": false,
        "description": "Price and greeks along spot (at several horizons) or along time (at several spot shifts).",
        "properties": {
          "axis": {
            "enum": [
              "spot",
              "time"
            ],
            "title": "Axis",
            "type": "string"
          },
          "greeks": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Greek"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Greeks to compute (default all; [] for value only)",
            "title": "Greeks"
          },
          "horizons_days": {
            "default": [
              0
            ],
            "description": "Spot axis: valuation-date rolls",
            "items": {
              "type": "integer"
            },
            "maxItems": 6,
            "minItems": 1,
            "title": "Horizons Days",
            "type": "array"
          },
          "points": {
            "default": 81,
            "maximum": 201.0,
            "minimum": 2.0,
            "title": "Points",
            "type": "integer"
          },
          "pricing": {
            "$ref": "#/components/schemas/PriceRequest"
          },
          "spot_range_pct": {
            "default": 30.0,
            "description": "Spot axis: \u00b1range, %",
            "exclusiveMaximum": 100.0,
            "exclusiveMinimum": 0.0,
            "title": "Spot Range Pct",
            "type": "number"
          },
          "spot_shifts_pct": {
            "default": [
              0.0
            ],
            "description": "Time axis: spot shifts, %",
            "items": {
              "type": "number"
            },
            "maxItems": 6,
            "minItems": 1,
            "title": "Spot Shifts Pct",
            "type": "array"
          }
        },
        "required": [
          "pricing",
          "axis"
        ],
        "title": "ProfileRequest",
        "type": "object"
      },
      "ProfileResponse": {
        "additionalProperties": false,
        "description": "Profiles along spot or time.",
        "properties": {
          "axis": {
            "enum": [
              "spot",
              "time"
            ],
            "title": "Axis",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "notional": {
            "description": "quantity \u00d7 current spot, currency",
            "title": "Notional",
            "type": "number"
          },
          "payoff": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Spot axis: position payoff at expiry",
            "title": "Payoff"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/ProfileSeries"
            },
            "title": "Series",
            "type": "array"
          },
          "spot": {
            "title": "Spot",
            "type": "number"
          },
          "strike": {
            "title": "Strike",
            "type": "number"
          },
          "units": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "string"
              },
              "propertyNames": {
                "$ref": "#/components/schemas/Greek"
              },
              "type": "object"
            },
            "propertyNames": {
              "enum": [
                "pure",
                "cash"
              ]
            },
            "title": "Units",
            "type": "object"
          },
          "x": {
            "description": "Spot level, or days to expiry",
            "items": {
              "type": "number"
            },
            "title": "X",
            "type": "array"
          }
        },
        "required": [
          "axis",
          "x",
          "currency",
          "notional",
          "strike",
          "spot",
          "payoff",
          "units",
          "series"
        ],
        "title": "ProfileResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Quant Pricer API",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/compare": {
      "post": {
        "description": "The product under every registered method, with convergence and exercise boundary.",
        "operationId": "compare_api_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Compare"
      }
    },
...
    "/api/heatmap": {
      "post": {
        "description": "Full-revaluation position values over spot \u00d7 vol shocks, optionally after a time roll.",
        "operationId": "heatmap_api_heatmap_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeatmapRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeatmapResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Heatmap"
      }
    },
    "/api/implied-vol": {
      "post": {
        "description": "Flat implied volatility reproducing a unit price (Let's Be Rational), European only.",
        "operationId": "implied_vol_api_implied_vol_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImpliedVolRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpliedVolResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Implied Vol"
      }
    },
...
    "/api/price": {
      "post": {
        "description": "Price one instrument with greeks in desk units.",
        "operationId": "price_api_price_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Price"
      }
    },
    "/api/profile": {
      "post": {
        "description": "Position value and desk greeks along spot or time (full revaluation at each point).\n\nNumerical methods run at the scenario resolution (``settings.scenario``).",
        "operationId": "profile_api_profile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Profile"
      }
    }
  }
}